voice-agent/pyproject.toml
[project]
name = "voice-agent-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[deepgram,openai,silero,turn-detector,cartesia,assemblyai]>=1.2.15",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
    "httpx>=0.27.0",
]
